requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1; python_version < '3.9'
django==4.2
djangorestframework==3.14.0
gunicorn==20.1.0
pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
